working rpeort

+++ text.h
@@ -1,3 +1,3 @@
 void search_and_replace(char *filepath);
 extern int occurence_found;
-extern int replacements_done;
+extern long int replacements_done;
